requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
exceptiongroup==1.3.1
    # via anyio
fastapi==0.128.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
loguru==0.7.3
    # via -r requirements.in
pillow==12.1.0
    # via -r requirements.in
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
segno==1.6.6
    # via -r requirements.in
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via -r requirements.in
win32-setctime==1.2.0
    # via loguru
